submission runs, adding notes
the submission runs as expected and as documented, great job!!
this is a very cool project. ive only tried the baseline so far,
but this seems like it will be fun to play around with a bit.

had to take some notes, because getting things to work on my system
had some challenges. they were simple enough to resolve, though,
so i've documented the changes i needed to make.

great project!